servidor/backend/requirements.txt
bcrypt==3.1.7
cffi==1.13.2
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
paho-mqtt==1.5.0
pkg-resources==0.0.0
pycparser==2.19
six==1.13.0
SQLAlchemy==1.3.11
Werkzeug==0.16.0
